refactor the code of service
